+++ ImageFeed/Configuration/HttpMethods.swift
@@ -1,0 +1,13 @@
+//
+//  HttpMethods.swift
+//  ImageFeed
+//
+//  Created by Dmitriy Noise on 16.05.2025.
+//
+
+enum HttpMethods {
+    static let get = "GET"
+    static let post = "POST"
+    static let head = "HEAD"
+    static let delete = "DELETE"
+}
